services/producer/requirements.txt
boto3==1.35.36
kafka-python==2.0.2
